functions/nsip_initial_example/requirements.txt
pydantic
azure-identity
azure-functions
azure-servicebus
azure-storage-blob